connection pool and max overflow added

+++ database/models/users/users.py
@@ -1,10 +1,10 @@
 from core.database.connection import Base
-from sqlalchemy import String, BigInteger
+from sqlalchemy import String, BigInteger, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
 class User(Base):
     __tablename__ = "users"
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     first_name: Mapped[str] = mapped_column(String(50), nullable=True)
     last_name: Mapped[str] = mapped_column(String(50), nullable=True)
     email: Mapped[str] = mapped_column(String(255), nullable=False)
